6 errors left dealing with 0 percision not printing a string containing a 0

+++ libft/ft_itoa.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char			*ft_itoa(long long n)
+char			*ft_itoa(intmax_t n)
 {
 	char		*str;
-	long long	num;
+	intmax_t	num;
 	size_t		len;
 
 	len = ft_count_int_len(n, 10);
